backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.11.1
attrs==25.4.0
crypto_cpp_py==1.4.5
cytoolz==1.1.0
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
ecdsa==0.18.0
eth-hash==0.7.1
eth-keyfile==0.9.1
eth-keys==0.7.0
eth-typing==5.2.1
eth-utils==5.3.1
frozenlist==1.8.0
gunicorn==25.1.0
idna==3.11
lark==1.3.1
marshmallow==3.26.2
marshmallow-oneofschema==3.2.0
marshmallow_dataclass==8.7.1
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
packaging==26.0
poseidon_py==0.1.5
propcache==0.4.1
psycopg2-binary==2.9.11
py-ecc==8.0.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
semver==3.0.4
six==1.17.0
sqlparse==0.5.5
starknet-py==0.29.0
sympy==1.12.1
toolz==1.1.0
typeguard==4.5.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
websockets==15.0.1
yarl==1.22.0
